requirements.txt
aioquic==1.2.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fake-useragent==2.1.0
fcm-django==2.2.1
filelock==3.16.1
firebase-admin==6.6.0
Flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==11.1.3
mitmproxy-windows==0.11.5
mitmproxy_rs==0.11.5
msgpack==1.1.0
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.6
playwright==1.51.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydivert==2.1.0
pyee==12.1.1
PyJWT==2.10.1
pylsqpack==0.3.19
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-telegram-bot==22.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.10
scapy==2.6.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.2
swapper==1.4.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
urwid==2.6.16
virtualenv==20.28.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
